+++ tools/proto/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 🐞 Fixes
 
+- Filter out `version_spec` tags from versions list.
 - Don't link bins/shims as it collides with the default install.
 
 ## 0.1.0